thêm thông tin dân tộc

+++ app/Http/Controllers/PhanQuyenController.php
@@ -24,44 +24,69 @@ class PhanQuyenController extends Controller
   }
   public function phanquyen(){
     $this->check_login();
-    $list_vienchuc = VienChuc::where('status_vc', '<>', '1')
-      ->orderBy('vienchuc.ma_vc', 'desc')
-      ->get();
-    $list_phanquyen = PhanQuyen::join('quyen', 'quyen.ma_q', '=', 'phanquyen.ma_q')
-      ->join('vienchuc', 'vienchuc.ma_vc', '=', 'phanquyen.ma_vc')
-      ->get();
-    $list_quyen = Quyen::where('status_q', '<>', '1')
-      ->orderBy('ma_q', 'desc')
-      ->get();
     $ma_vc = session()->get('ma_vc');
     $phanquyen_admin = PhanQuyen::where('ma_vc', $ma_vc)
       ->where('ma_q', '=', '5')
       ->first();
-    return view('quyen.phanquyen')
-      ->with('list_quyen', $list_quyen)
-      ->with('phanquyen_admin', $phanquyen_admin)
-      ->with('list_phanquyen', $list_phanquyen)
-      ->with('list_vienchuc', $list_vienchuc);
-    
+    if($phanquyen_admin){
+      $list_vienchuc = VienChuc::join('khoa', 'khoa.ma_k', '=', 'vienchuc.ma_k')
+        ->where('status_vc', '<>', '1')
+        ->orderBy('vienchuc.ma_vc', 'desc')
+        ->get();
+      $list_phanquyen = PhanQuyen::join('quyen', 'quyen.ma_q', '=', 'phanquyen.ma_q')
+        ->join('vienchuc', 'vienchuc.ma_vc', '=', 'phanquyen.ma_vc')
+        ->get();
+      $list_quyen = Quyen::where('status_q', '<>', '1')
+        ->orderBy('ma_q', 'desc')
+        ->get();
+      $ma_vc = session()->get('ma_vc');
+      $phanquyen_admin = PhanQuyen::where('ma_vc', $ma_vc)
+        ->where('ma_q', '=', '5')
+        ->first();
+      return view('quyen.phanquyen')
+        ->with('list_quyen', $list_quyen)
+        ->with('phanquyen_admin', $phanquyen_admin)
+        ->with('list_phanquyen', $list_phanquyen)
+        ->with('list_vienchuc', $list_vienchuc);
+    }else{
+      return Redirect::to('/home');
+    }
   }
   public function phanquyen_vc(Request $request){
     $this->check_login();
-    $data = $request->all();
-    $phanquyen = new PhanQuyen();
-    $phanquyen->ma_vc = $data['ma_vc'];
-    $phanquyen->ma_q = $data['ma_q'];
-    $phanquyen->save();
-    $request->session()->put('message','Thêm thành công');
-    return Redirect::to('/phanquyen');
+    $ma_vc = session()->get('ma_vc');
+    $phanquyen_admin = PhanQuyen::where('ma_vc', $ma_vc)
+      ->where('ma_q', '=', '5')
+      ->first();
+    if($phanquyen_admin){
+      $data = $request->all();
+      $phanquyen = new PhanQuyen();
+      $phanquyen->ma_vc = $data['ma_vc'];
+      $phanquyen->ma_q = $data['ma_q'];
+      $phanquyen->save();
+      $request->session()->put('message','Thêm thành công');
+      return Redirect::to('/phanquyen');
+    }else{
+      return Redirect::to('/home');
+    }
+    
   }
   public function lammoi_quyen($ma_vc){
     $this->check_login();
-    $list_phanquyen_ma = PhanQuyen::where('ma_vc', $ma_vc)
-      ->where('ma_q', '<>', '10')
-      ->get();
-    foreach($list_phanquyen_ma as  $phanquyen){
-      $phanquyen->delete();
+    $ma_vc = session()->get('ma_vc');
+    $phanquyen_admin = PhanQuyen::where('ma_vc', $ma_vc)
+      ->where('ma_q', '=', '5')
+      ->first();
+    if($phanquyen_admin){
+      $list_phanquyen_ma = PhanQuyen::where('ma_vc', $ma_vc)
+        ->where('ma_q', '<>', '10')
+        ->get();
+      foreach($list_phanquyen_ma as  $phanquyen){
+        $phanquyen->delete();
+      }
+      return Redirect::to('/phanquyen');
+    }else{
+      return Redirect::to('/home');
     }
-    return Redirect::to('/phanquyen');
   }
 }
